+++ src/franz/openrdf/model/literal.py
@@ -220,6 +220,24 @@ class Literal(Value):
             sb.append(self.datatype.toNTriples())
         return ''.join(sb)
 
+    def to_json_ld(self):
+        """ Converts to an object to be used as a JSON-LD value. """
+        if self.language:
+            return {
+                '@value': self.label,
+                '@language': self.language
+            }
+
+        if self.datatype is None or self.datatype == XMLSchema.STRING:
+            return self.label
+
+        if self.datatype == XMLSchema.INTEGER:
+            return int(self.label)
+
+        return {
+            '@value': self.label,
+            '@type': self.datatype.to_json_ld_key()
+        }
 
 ###############################################################################
 ## Automatic conversion from Literal to Python object